multi_agent_qa_api/requirements.txt

fastapi[all]
uvicorn
langchain
langchain-community
langchain-core
langchain-openai
sentence-transformers
faiss-cpu
python-dotenv
pypdf
jinja2  